requirements.txt
# Core dependencies
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# PDF Processing
PyPDF2==3.0.1
pdfplumber==0.10.3
pypdf==3.17.4
pymupdf==1.23.8

# Text Processing & NLP
spacy==3.7.2
nltk==3.8.1
textract==1.6.5
python-docx==1.1.0

# Data Processing
pandas==2.1.4
numpy==1.26.3
openpyxl==3.1.2

# API & Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
httpx==0.26.0

# AI & Machine Learning
openai==1.51.2

# Testing
pytest==7.4.4
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Utilities
rich==13.7.0
click==8.1.7
tqdm==4.66.1
schedule==1.2.0

# Monitoring & Logging
loguru==0.7.2
sentry-sdk==1.39.1